move image library button from dashboard to topbar
