+++ packages/keystatic/src/app/dashboard/DashboardPage.tsx
@@ -2,6 +2,8 @@ import { useLocalizedStringFormatter } from '@react-aria/i18n';
 
 import { Avatar } from '@keystar/ui/avatar';
 import { Flex, VStack } from '@keystar/ui/layout';
+import { TextLink } from '@keystar/ui/link';
+import { tokenSchema } from '@keystar/ui/style';
 import { Heading } from '@keystar/ui/typography';
 
 import { Config } from '../../config';
@@ -11,10 +13,7 @@ import { useViewer } from '../shell/viewer-data';
 import { BranchSection } from './BranchSection';
 import { CollectionSection } from './CollectionSection';
 import { SingletonSection } from './SingletonSection';
-import { tokenSchema } from '@keystar/ui/style';
-import { ImagesSection } from './ImagesSection';
 import { useCloudInfo } from '../shell/data';
-import { TextLink } from '@keystar/ui/link';
 
 export function DashboardPage(props: { config: Config; basePath: string }) {
   const stringFormatter = useLocalizedStringFormatter(l10nMessages);
@@ -37,7 +36,6 @@ export function DashboardPage(props: { config: Config; basePath: string }) {
           {user && <UserInfo user={user} manageAccount={!!cloudInfo} />}
 
           <BranchSection config={props.config} />
-          <ImagesSection />
           <CollectionSection basePath={props.basePath} config={props.config} />
           <SingletonSection basePath={props.basePath} config={props.config} />
         </Flex>
